added files test for all the classes
