--- src/main/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/interfaces/Heater.java
@@ -1,6 +1,0 @@
-package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.interfaces;
-
-public interface Heater {
-    void on();
-    void off();
-}
