--- src/main/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/interfaces/Sensor.java
@@ -1,5 +1,0 @@
-package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.interfaces;
-
-public interface Sensor {
-    boolean getState();
-}
