+++ src/main/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testImplementation/CoffeeMaker.java
@@ -1,12 +1,12 @@
-package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.implementation;
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testImplementation;
 
-import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.boiler.Boiler;
-import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.pieces.DeliveryPipe;
-import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.pieces.FilterAndReceptacle;
-import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.pieces.IndicatorLight;
-import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.pieces.StartButton;
-import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.pot.PlateSensor;
-import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.pot.Pot;
+import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testBoiler.Boiler;
+import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPieces.DeliveryPipe;
+import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPieces.FilterAndReceptacle;
+import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPieces.IndicatorLight;
+import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPieces.StartButton;
+import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPot.PlateSensor;
+import org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPot.Pot;
 
 public class CoffeeMaker {
     private Boiler boiler;
