+++ src/main/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testImplementation/Inputs.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.implementation;
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testImplementation;
 
 import java.util.Scanner;
