+++ src/main/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testInterfaces/Container.java
@@ -1,0 +1,5 @@
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testInterfaces;
+
+public interface Container {
+    boolean containLiquid();
+}
