+++ src/main/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testInterfaces/Heater.java
@@ -1,0 +1,6 @@
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testInterfaces;
+
+public interface Heater {
+    void on();
+    void off();
+}
