+++ src/main/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testPieces/FilterAndReceptacle.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.pieces;
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPieces;
 
 public class FilterAndReceptacle {
     private boolean coffeeGrains;
