+++ src/main/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testPieces/IndicatorLight.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.pieces;
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPieces;
 
 public class IndicatorLight {
     private boolean lightState;
