+++ src/test/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testBoiler/BoilerTest.java
@@ -1,0 +1,12 @@
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testBoiler;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class BoilerTest {
+
+    @Test
+    public void containLiquid() {
+    }
+}
