+++ src/test/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testImplementation/InputsTest.java
@@ -1,0 +1,12 @@
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testImplementation;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class InputsTest {
+
+    @Test
+    public void scanner() {
+    }
+}
