+++ src/test/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testPieces/DeliveryPipeTest.java
@@ -1,0 +1,12 @@
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPieces;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class DeliveryPipeTest {
+
+    @Test
+    public void getDeliveryWarmWater() {
+    }
+}
