+++ src/test/java/org/fundacionjala/prog101CoffeeMaker/coffeeMaker/testPot/PotTest.java
@@ -1,0 +1,28 @@
+package org.fundacionjala.prog101CoffeeMaker.coffeeMaker.testPot;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class PotTest {
+
+    @Test
+    public void containLiquid() {
+    }
+
+    @Test
+    public void getIsInPlace() {
+    }
+
+    @Test
+    public void setIsInPlace() {
+    }
+
+    @Test
+    public void getAmountOfCups() {
+    }
+
+    @Test
+    public void setAmountOfCups() {
+    }
+}
